+++ CHANGELOG.md
@@ -12,7 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Mirage
-    - `osfNestedResource` - added custom `post` handler to fix `create` action
+    - `osfNestedResource`
+        - added custom `post` handler to fix `create` action
+    - `node/contributors` nested resource
+        - conditionally create bibliographic contributors when creating contributors
 
 ## [19.8.0] - 2019-08-15
 ### Added